+++ go.mod
@@ -7,6 +7,7 @@ require golang.zx2c4.com/wireguard/wgctrl v0.0.0-20200324154536-ceff61240acf
 require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/golang/protobuf v1.3.5
+	github.com/square/certstrap v1.2.0 // indirect
 	golang.org/x/crypto v0.0.0-20200204104054-c9f3fb736b72
 	golang.org/x/oauth2 v0.0.0-20180821212333-d2e6202438be
 	google.golang.org/grpc v1.28.1
