+++ server.go
@@ -2,16 +2,8 @@ package wgrpcd
 
 import (
 	"context"
-	"crypto/tls"
-	"crypto/x509"
-	"fmt"
-	"io/ioutil"
-	"log"
-	"net/http"
 	"os"
-	"path/filepath"
 
-	"golang.org/x/crypto/acme/autocert"
 	"golang.zx2c4.com/wireguard/wgctrl/wgtypes"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
@@ -246,40 +238,10 @@ func (s *Server) authResult(ctx context.Context) *AuthResult {
 }
 
 // NewServer returns a wgrpcd instance configured to use a gRPC server with TLSv1.3.
-// It will automatically fetch a SSL certificate from Let's Encrypt.
-// wgrpcd refuses all unencrypted connections.
 func NewServer(config *ServerConfig) (*grpc.Server, error) {
-	certManager := autocert.Manager{
-		Prompt:     autocert.AcceptTOS,
-		HostPolicy: autocert.HostWhitelist(config.Hostname),
-		Cache:      autocert.DirCache(cacheDir(config.Hostname)),
-	}
-
-	go http.ListenAndServe(":http", certManager.HTTPHandler(nil))
-
-	// Load the CA certificate
-	trustedCert, err := ioutil.ReadFile(config.CACertFilename)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load trusted certificate: %w", err)
-	}
-
-	// Put the CA certificate to certificate pool
-	certPool := x509.NewCertPool()
-	if !certPool.AppendCertsFromPEM(trustedCert) {
-		return nil, fmt.Errorf("failed to append trusted certificate to certificate pool: %w", err)
-	}
-
-	// Create the TLS configuration
-	// Since this is gRPC, we can enforce TLSv1.3.
-	tlsConfig := &tls.Config{
-		GetCertificate: certManager.GetCertificate,
-		ClientCAs:      certPool,
-		MinVersion:     tls.VersionTLS13,
-		MaxVersion:     tls.VersionTLS13,
-	}
 
 	// Create a new TLS credentials based on the TLS configuration and return a gRPC server configured with this.
-	cred := credentials.NewTLS(tlsConfig)
+	cred := credentials.NewTLS(config.TLSConfig)
 	authority := &Authority{Logger: config.Logger}
 
 	if config.AuthProvider != nil {
@@ -295,17 +257,4 @@ func NewServer(config *ServerConfig) (*grpc.Server, error) {
 	)
 	RegisterWireguardRPCServer(rpcServer, &Server{logger: config.Logger})
 	return rpcServer, nil
-}
-
-func cacheDir(hostname string) (dir string) {
-	dir = filepath.Join(os.TempDir(), "cache-golang-autocert-"+hostname)
-	if _, err := os.Stat(dir); !os.IsNotExist(err) {
-		log.Println("Found cache dir:", dir)
-		return dir
-	}
-	if err := os.MkdirAll(dir, 0700); err == nil {
-		return dir
-	}
-
-	panic("couldnt create cert cache directory")
 }
